refactor: isolate avatar controller
