+++ src/server/controllers/index.js
@@ -1,5 +1,7 @@
 import * as collectives from './collectives';
+import avatar from './avatar';
 
 export default {
   collectives,
+  avatar,
 };
